Add arrows between roles in header
Because of some *strange* opacity issues while positioning, these arrows
are placed as additional list items after the roles themselves. This
isn't great from an HTML-perspective, but it makes styling easier, and
try as I may I could not figure out a way to prevent opacity issues when
placing them inside the main list items.

When positioning them anywhere outside the main container (e.g. with
negative positioning or margins), it would cause whatever container it
was covering to have double-opacity, from what I could tell. Maybe this
is a paint issue where if something with transparency (even if it's
small) lays over another element with transparency, the top element
causes it to be transparent twice?

One thing we could try is to set the stroke color of the arrow to have
an alpha color, instead of setting the opacity of the container.
But I'm tired at this moment and maybe I can try that later. It's
working now though!